+++ doc/change-log.md
@@ -15,6 +15,7 @@
 - 引入 `com.lucendar.common.db.jdbc.DbHelper.timestampStmtSetter`
 - 引入 `com.lucendar.common.db.jdbc.ResultSetAccessor.date()`, `timestamp()` 方法
 - 引入 `com.lucendar.common.db.jdbc.StatementBinder.setDate` 方法
+- 引入 `com.lucendar.common.db.jdbc.DbHelper.dateStmtSetter` 方法
 
 ## 变更
 - 不兼容变更